LPD-28681 Rename definitions.ts file to event-definitions.ts
