+++ modules/test/playwright/tests/osb-faro-web/utils/event-definitions.ts
